UPDATE: cleanup updates (v11.18.0)
- new cleanup option `--downloads` to remove downloads caches (conda, pip,
  and templates)
- change: now conda pkgs is cleaned up also in quick cleanup (which now
  includes all "downloads" cleanups)
- robot cache is now part of full cleanup
- run commands now cleanup their temp folders immediately

+++ cmd/rcc/main.go
@@ -80,10 +80,13 @@ func markTempForRecycling() {
 	if markedAlready {
 		return
 	}
-	markedAlready = true
-	filename := filepath.Join(common.RobocorpTemp(), "recycle.now")
-	ioutil.WriteFile(filename, []byte("True"), 0o644)
-	common.Debug("Marked %q for recycling.", common.RobocorpTemp())
+	target := common.RobocorpTempName()
+	if pathlib.Exists(target) {
+		filename := filepath.Join(target, "recycle.now")
+		ioutil.WriteFile(filename, []byte("True"), 0o644)
+		common.Debug("Marked %q for recycling.", target)
+		markedAlready = true
+	}
 }
 
 func main() {
